[Vim] Use <Esc> to switch from terminal mode to normal mode
If you actually want to send an Escape key, use <C-v><Esc>.

I had tried <C-\><C-\> for going to normal mode, but this was an issue
outside Vim, it kills the running console. Because of muscle memory, I
accidentally closed the running console.

<Esc> feels natural, as it's the same with Insert mode.

`<C-v><Esc>` is trickier, but it's actually what `<C-v>` was designed
for (see Verbatim).